Start shuffling page elements around
Based on Michael's feedback
