+++ webpack/pages/block_page/index.js
@@ -6,7 +6,6 @@ import { graphql } from "react-apollo"
 import BlockPageQuery from "graphql/block_page.gql"
 
 import BlockSource from "./block_source";
-import BlockUsage from "./block_usage";
 import BlockRuns from "./block_runs";
 import Subscription from "./subscription"
 import Title from "./title"
@@ -29,28 +28,16 @@ const BlockPage = ({ data }) => (
           user={data.block.author}
         />
 
-        { data.session
-        ? <BlockUsage
-            editable={data.block.editable}
-            initial_input_data={data.block.initial_input_data}
-            name={data.block.name}
-            run_block_url={`${window.location.href}/runs.json`}
-            schema={data.block.schema}
-            user={data.block.author}
-            user_api_key={data.session.api_key}
-          />
-        : <Section>
-            Sign in with GitHub to try out this block.
-          </Section>
-        }
-
         <BlockSource
           id={data.block.id}
           environment={data.block.environment}
           editable={data.block.editable}
           name={data.block.name}
           source={data.block.source}
-          user={data.block.author}
+          session={data.session}
+          initial_input_data={data.block.initial_input_data}
+          schema={data.block.schema}
+          user_api_key={data.session.api_key}
         />
 
         <Subscription
